add student

+++ studentmanagementsystem/src/main/java/com/studentmanagementsystem/service/ClassesService.java
@@ -1,8 +1,11 @@
 package com.studentmanagementsystem.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.studentmanagementsystem.entity.Classes;
@@ -14,22 +17,29 @@ public class ClassesService {
 	@Autowired
 	private ClassRepository classRepository;
 
-	public Classes addClasses(Classes entity) {
-		
-		return classRepository.save(entity);
+	public ResponseEntity<String> addClasses(Classes entity) {
+
+		try {
+			classRepository.save(entity);
+			return new ResponseEntity<>("success", HttpStatus.CREATED);
+		} catch (Exception e) {
+			return new ResponseEntity<>("failure", HttpStatus.BAD_REQUEST);
+		}
 	}
 
-	public Classes getClass(int id) {
-		
-		return classRepository.findById(id).orElseThrow(()->new RecordNotFoundException("No Class Present in this id "+id));
+	public Classes getClass(Long id) {
+
+		return classRepository.findById(id)
+				.orElseThrow(() -> new RecordNotFoundException("No Class Present in this id " + id));
 	}
 
 	public List<Classes> getClasses() {
 		return classRepository.findAll();
 	}
 
-	public Classes updateClass(int id, Classes update) {
-		Classes findClass=classRepository.findById(id).orElseThrow((()->new RecordNotFoundException("No Class Present in this id "+id)));
+	public Classes updateClass(Long id, Classes update) {
+		Classes findClass = classRepository.findById(id)
+				.orElseThrow((() -> new RecordNotFoundException("No Class Present in this id " + id)));
 		findClass.setSection(update.getSection());
 		findClass.setClasses(update.getClasses());
 		findClass.setSubject1(update.getSubject1());
@@ -40,8 +50,13 @@ public class ClassesService {
 		return classRepository.save(findClass);
 	}
 
-	public void deleteClass(int id) {
-		classRepository.deleteById(id);
+	public void deleteClass(Long id) {
+		Classes findClass = classRepository.findById(id)
+				.orElseThrow((() -> new RecordNotFoundException("No Class Present in this id " + id)));
+		if (findClass != null) {
+			classRepository.deleteById(id);
+		}
+
 	}
 
 }
